+++ src/lib/axios/customAxios.ts
@@ -1,8 +1,6 @@
-import axios, { AxiosError } from 'axios';
-import { getRefreshToken } from 'lib/token/token';
-import { AccessExpression } from 'typescript';
+import axios from 'axios';
+import { tokenExpired } from 'lib/token/tokenExpired';
 
-// 인증이 필요 없는
 const customAxios = axios.create({
   baseURL: process.env.REACT_APP_BASE_URL,
   timeout: 10000,
@@ -12,10 +10,10 @@ const customAxios = axios.create({
 });
 
 customAxios.interceptors.request.use(
-  function (config) {
+  (config) => {
     return config;
   },
-  function (error) {
+  (error) => {
     return Promise.reject(error);
   },
 );
@@ -25,28 +23,14 @@ customAxios.interceptors.response.use(
     console.log(response);
     return response;
   },
-  async (error: AxiosError) => {
+  (error) => {
     if (error.response) {
       if (error.response?.status === 401) {
-        // 토큰 재발급 로직
-        const refreshToken = getRefreshToken();
-        try {
-          const { data } = await customAxios.put('/auth', null, {
-            headers: {
-              'Refresh-Token': `${refreshToken}`,
-            },
-          });
-          localStorage.setItem('access-token', data.accessToken);
-          alert('다시 시도해 주세요 !');
-        } catch (err) {
-          console.log(err);
-        }
+        tokenExpired();
       }
+      alert(error.request.response);
     }
-    alert(error.request.response);
-    // 에러 메세지 나중에 고치자...
-    // 심청이 하고 와서 고치자..
-    return error;
+    return Promise.reject(error);
   },
 );
 export { customAxios };
